# src/EasyIotSharp.API/StaticFiles/swagger/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "service.easyiotsharp：项目服务",
    "version": "v1"
  },
  "paths": {
...
    "/Tenant/Get": {
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "通过租户id获取一条租户信息",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDtoUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDtoUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDtoUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Tenant/Query": {
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "通过条件分页查询租户列表",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/QueryTenantInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryTenantInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryTenantInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QueryTenantInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDtoPagedResultDtoUPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDtoPagedResultDtoUPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantDtoPagedResultDtoUPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Tenant/Insert": {
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "添加一条租户信息",
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTenantInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTenantInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTenantInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTenantInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Tenant/Update": {
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "通过id修改一个租户信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Tenant/UpdateFreeze": {
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "通过id修改一个租户的冻结状态",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsFreezeTenantTenantInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsFreezeTenantTenantInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsFreezeTenantTenantInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsFreezeTenantTenantInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Tenant/Delete": {
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "通过id删除一个租户信息",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteTenantInput"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteTenantInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteTenantInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteTenantInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UPrimeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TenantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "creationTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "isDelete": {
            "type": "boolean",
            "description": "是否删除"
          },
          "deleteTime": {
            "type": "string",
            "description": "删除时间",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "description": "修改时间",
            "format": "date-time"
          },
          "operatorId": {
            "type": "string",
            "description": "操作人标识",
            "nullable": true
          },
          "operatorName": {
            "type": "string",
            "description": "操作人",
            "nullable": true
          },
          "numId": {
            "type": "integer",
            "description": "租户numId",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "租户名称",
            "nullable": true
          },
          "storeKey": {
            "type": "string",
            "description": "机构的（唯一的等同Id）",
            "nullable": true
          },
          "contractName": {
            "type": "string",
            "description": "合同名称",
            "nullable": true
          },
          "contractOwnerName": {
            "type": "string",
            "description": "合同所属人",
            "nullable": true
          },
          "contractOwnerMobile": {
            "type": "string",
            "description": "合同所属人电话",
            "nullable": true
          },
          "contractStartTime": {
            "type": "string",
            "description": "合同开始时间",
            "format": "date-time"
          },
          "contractEndTime": {
            "type": "string",
            "description": "合同结束时间",
            "format": "date-time"
          },
          "mobile": {
            "type": "string",
            "description": "机构的负责人手机号",
            "nullable": true
          },
          "owner": {
            "type": "string",
            "description": "机构的负责人姓名",
            "nullable": true
          },
          "storeLogoUrl": {
            "type": "string",
            "description": "机构的Logo url",
            "nullable": true
          },
          "remark": {
            "type": "string",
            "description": "机构的备注",
            "nullable": true
          },
          "email": {
            "type": "string",
            "description": "机构的电子邮箱",
            "nullable": true
          },
          "expiredType": {
            "type": "integer",
            "description": "过期状态\r\n0=待授权\r\n1=生效\r\n2=已过期",
            "format": "int32",
            "readOnly": true
          },
          "managerId": {
            "type": "string",
            "description": "机构的系统管理员Id 详见 EasyIotSharp.Core.Domain.TenantAccount.Soldier",
            "nullable": true
          },
          "versionTypeId": {
            "type": "integer",
            "description": "版本类型，详见 Elasticsearch.Net.VersionType\r\n1=入门版\r\n2=基础版\r\n3=旗舰版\r\n5=全国版",
            "format": "int32"
          },
          "isFreeze": {
            "type": "boolean",
            "description": "是否冻结"
          },
          "freezeDes": {
            "type": "string",
            "description": "冻结描述",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "代表一个租户实体的“DTO”"
      },
      "TenantDtoUPrimeResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/TenantDto"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "fullMessage": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "QueryTenantInput": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string",
            "description": "租户名称",
            "nullable": true
          },
          "expiredType": {
            "type": "integer",
            "description": "过期状态\r\n-1=不参与查询\r\n0=待授权\r\n1=生效\r\n2=已过期",
            "format": "int32"
          },
          "contractStartTime": {
            "type": "string",
            "description": "合同开始时间",
            "format": "date-time",
            "nullable": true
          },
          "contractEndTime": {
            "type": "string",
            "description": "合同结束时间",
            "format": "date-time",
            "nullable": true
          },
          "isFreeze": {
            "type": "integer",
            "description": "冻结状态\r\n-1=不参与查询\r\n1=已冻结\r\n2=未冻结",
            "format": "int32"
          },
          "pageIndex": {
            "type": "integer",
            "description": "当前开始页",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "页大小",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "通过条件分页查询租户列表的入参类"
      },
      "TenantDtoPagedResultDto": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TenantDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TenantDtoPagedResultDtoUPrimeResponse": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/TenantDtoPagedResultDto"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "fullMessage": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "InsertTenantInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "租户名称",
            "nullable": true
          },
          "contractName": {
            "type": "string",
            "description": "合同名称",
            "nullable": true
          },
          "contractOwnerName": {
            "type": "string",
            "description": "合同所属人",
            "nullable": true
          },
          "contractOwnerMobile": {
            "type": "string",
            "description": "合同所属人电话",
            "nullable": true
          },
          "contractStartTime": {
            "type": "string",
            "description": "合同开始时间",
            "format": "date-time"
          },
          "contractEndTime": {
            "type": "string",
            "description": "合同结束时间",
            "format": "date-time"
          },
          "mobile": {
            "type": "string",
            "description": "机构的负责人手机号",
            "nullable": true
          },
          "owner": {
            "type": "string",
            "description": "机构的负责人姓名",
            "nullable": true
          },
          "storeLogoUrl": {
            "type": "string",
            "description": "机构的Logo url",
            "nullable": true
          },
          "remark": {
            "type": "string",
            "description": "机构的备注",
            "nullable": true
          },
          "email": {
            "type": "string",
            "description": "机构的电子邮箱",
            "nullable": true
          },
          "versionTypeId": {
            "type": "integer",
            "description": "版本类型，详见 Elasticsearch.Net.VersionType\r\n1=入门版\r\n2=基础版\r\n3=旗舰版\r\n5=全国版",
            "format": "int32"
          },
          "isFreeze": {
            "type": "boolean",
            "description": "是否冻结"
          },
          "freezeDes": {
            "type": "string",
            "description": "冻结描述",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UPrimeResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "fullMessage": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "UpdateTenantInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "租户id",
            "nullable": true
          },
          "numId": {
            "type": "integer",
            "description": "租户numId",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "租户名称",
            "nullable": true
          },
          "storeKey": {
            "type": "string",
            "description": "机构的（唯一的等同Id）",
            "nullable": true
          },
          "contractName": {
            "type": "string",
            "description": "合同名称",
            "nullable": true
          },
          "contractOwnerName": {
            "type": "string",
            "description": "合同所属人",
            "nullable": true
          },
          "contractOwnerMobile": {
            "type": "string",
            "description": "合同所属人电话",
            "nullable": true
          },
          "contractStartTime": {
            "type": "string",
            "description": "合同开始时间",
            "format": "date-time"
          },
          "contractEndTime": {
            "type": "string",
            "description": "合同结束时间",
            "format": "date-time"
          },
          "mobile": {
            "type": "string",
            "description": "机构的负责人手机号",
            "nullable": true
          },
          "storeLogoUrl": {
            "type": "string",
            "description": "机构的Logo url",
            "nullable": true
          },
          "remark": {
            "type": "string",
            "description": "机构的备注",
            "nullable": true
          },
          "email": {
            "type": "string",
            "description": "机构的电子邮箱",
            "nullable": true
          },
          "managerId": {
            "type": "integer",
            "description": "机构的系统管理员Id 详见 EasyIotSharp.Core.Domain.TenantAccount.Soldier",
            "format": "int32"
          },
          "versionTypeId": {
            "type": "integer",
            "description": "版本类型，详见 Elasticsearch.Net.VersionType\r\n1=入门版\r\n2=基础版\r\n3=旗舰版\r\n5=全国版",
            "format": "int32"
          },
          "isFreeze": {
            "type": "boolean",
            "description": "是否冻结"
          },
          "freezeDes": {
            "type": "string",
            "description": "冻结描述",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "通过id修改一个租户信息的入参类"
      },
      "UpdateIsFreezeTenantTenantInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "租户id",
            "nullable": true
          },
          "isFreeze": {
            "type": "boolean",
            "description": "是否冻结"
          },
          "freezeDes": {
            "type": "string",
            "description": "冻结描述",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "通过id修改一个租户的冻结状态得入参类"
      },
      "DeleteTenantInput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "租户id",
            "nullable": true
          },
          "isDelete": {
            "type": "boolean",
            "description": "是否删除"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Value Bearer {token}",
        "name": "u-token",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}